% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paths_To_Match.R
\name{Paths_To_Match}
\alias{Paths_To_Match}
\title{Match Hallmark Paths to Core Enrichment Column}
\usage{
Paths_To_Match(x, y)
}
\arguments{
\item{x}{Matches Genes in the opened gmt object previously created and matches them to genes
found within the core_enrichment table of the enriched signatures table}

\item{y}{The name of the object containing all of the opened .gmt files}
}
\value{
A dataframe with a new column titled "matched_pathways". Whether the genes in y
is present in x, the name of the .gmt file in the y object will be attached in this column.
}
\description{
Match Hallmark Paths to Core Enrichment Column
}
\examples{
Paths_To_Match(df,open_dataset)
}
