% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset_To_Hallmarkpaths.R
\name{Subset_To_Hallmarkpaths}
\alias{Subset_To_Hallmarkpaths}
\title{Subset To Hallmark Pathways}
\usage{
Subset_To_Hallmarkpaths(x)
}
\arguments{
\item{x}{Subset Enriched Signatures Table to only Hallmark Paths}
}
\value{
A new dataframe of the same name that will return the Hallmark Paths
}
\description{
Subset To Hallmark Pathways
}
\examples{
Subset_To_Hallmarkpaths(x)
}
