% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_genesets.R
\name{meta_genesets}
\alias{meta_genesets}
\title{Meta Genesets}
\usage{
meta_genesets(x)
}
\arguments{
\item{x}{To use this function, all.gmt files downloaded from GSEA would
ideally be organized into a single folder.}
}
\value{
All file names within that folder will be assigned to a single object.
The open_genesets function will then read all of the files within that singular object
}
\description{
Meta Genesets
}
\examples{
meta_genesets <- function("/insert/path/to/folder/here")
}
