% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_genesets.R
\name{open_genesets}
\alias{open_genesets}
\title{Open Genesets}
\usage{
open_genesets(x)
}
\arguments{
\item{x}{This is where we open every file within the meta-file object we created using the
meta_genesets function}
}
\value{
All files within the object created by meta_genesets are read. Each row will not contain the genes associated with
the .gmt filenames from each row.
Object is in a list format.
}
\description{
Open Genesets
}
\examples{
open_genesets <- open_genesets(meta_genesets)
}
